[JENKINS-37087] Iterate over all matching templates
Currently, if multiple pod templates match the label expression of a
particulat job, pod is tried to be spawned from only the first
matching template. If, in the scenario, when the container cap for
the first matching template is reached, the job would starve, even
though there might be other pod templates which could be used. This
patch fixes this issue by iterating over all matching templates.