move transients into DecoratedLauncher
as `decorate` will be called multiple times, and
we don't want to share those among calls.

Makes JENKINS-40825 much less severe, as
"Pipe not connected" now most of the time seems to
come from [DurableTaskStep][1], which will just retry
instead of failing the job.

Note: this is just a first stab and needs polishing

[1]: https://github.com/jenkinsci/workflow-durable-task-step-plugin/blob/ae18393/src/main/java/org/jenkinsci/plugins/workflow/steps/durable_task/DurableTaskStep.java#L332