[JENKINS-35246] Move ProvisioningCallback logic to KubernetesLauncher.
This fixes the provisioning lifecycle for KubernetesSlave. The problem
was ProvisioningCallback was adding the node to Jenkins but it is
expected to be done by NodeProvisioner once the future completes.

This could lead in cases of short node usage in cases where the node would
get readded to Jenkins after the termination had occurred.

Moving the provisioning logic to a ComputerLauncher honors the lifecycle
properly since ComputerLauncher#launch is called after the
NodeProvisioner has finished its job, eliminating risk of race
condition.