Use single quotes to export variables.
That way, we only need to escape single quotes, nothing else.

Single quotes cannot be escaped using baslash, as that is treated
literally inside single quotes.
I.e. to get a literal single quotes, we have to break out of the
single quotes and escape it outside. "Simple".

So "it's cool" becomes 'it'\''s cool,i',

Treated by the shell as the concatenation of "it", "'", and "s cool".

See e.g. https://stackoverflow.com/a/1315213.