Declarative: do not wrap steps with `container` statement unless default container is changed
We met a weird issue when migrating the scripted syntax to declarative
with Kubernetes plugin: The `Launcher` doesn't forward `stdin`,
`stdout`, or `stderr` even though they are requested to be forwarded.
This breaks many plugins that use `Launcher` to start a remote process on Jenkins slaves.

This happens because when Kubernetes plugin converts declarative syntax to scripted syntax,
all build steps are enclosed with a `container` statement.
As a result, remote commands are run via the Kubernetes exec API rather than
the jnlp agent.

Since running steps via Kubernetes exec API is still considered ALPHA
and there are some compatibility issues to solve,
I would like to make a small change to the syntax converting process:
The declarative steps will not be enclosed with a `container` statement unless
an alternate default container is specified by the pipeline user.