Pod Labels UI
For clusters that run multiple jenkins instance in the same namespace
and also want to limit the number of slave pods by instance, being
able to manage the pod labels is important.

Because there was no UI and thus no DataBoundSetter any labels
set via an init.groovy.d script or script console would get wiped out
on any config change.

This changeset introduces a `labelsStr` property which is used to
bind the data from the UI form. The string format is a simple key=value
pair separated by whitespace. Do to the restricted syntax and character
set defined by kubernetes parsing of the labels is pretty straightforward.