JENKINS-58740: add missing pod label field attribute to config.jelly
The `repeatableHeteroProperty` does not get the field from the `entry` parent tag and results in a NPE when rendering the configure page. The `field` attribute is on both the `entry` and `repeatableHeteroProperty` to render the help correctly. This follows precedent set by other fields like `annotations` on PodTemplate.

fixes JENKINS-58740