readme: add instructions for running with OpenShift 3
Release 1.26.0 introduced an update [1], which made older versions of Kubernetes
fail rather obscurely, and with them, OpenShift 3 too. The instruction provided
in the issue discussion and release notes might be hard to find for new users
of the plugin, thus adding them to the readme.

As OpenShift 3.11 will still be maintained until June 2022 [2], many
users might be hitting this issue for a while still.

[1] https://github.com/jenkinsci/kubernetes-plugin/pull/790
[2] https://access.redhat.com/support/policy/updates/openshift_noncurrent