Avoid canceling build if resourcequota exceeded or conflict
A change introduced in v1.27.1 meant that queued build tasks will be
canceled and terminated with failure if Kubernetes returns 4xx error.

This change prevents failing the build if a resourcequota is exceeded or
in conflict. Instead, the queued item will continuously retry scheduling
until successful or aborted.

This is useful in envs which use resourcequotas and want the pod to
eventually schedule when some resource is free. Additionally on some
kubernetes clusters a 409 response may be returned if many pods are
scheduled in a short amount of time (typically parallel jobs), a
immediate retry typically resolve this.

Related ticket: JENKINS-63976