Update Reaper to manage watchers across multiple clouds and config updates
Update Reaper to keep track of watchers across multiple clouds and add/remove
them when clouds are updated. To better track when watchers need to be
recreated the KubernetesClientProvider#getValidity method was made public.
This allows the Reaper to validate the watch client connection is still valid if
the cloud connection properties are updated.

If a watcher is closed due to "410 Gone" responses the watcher will deregister
itself and reregister next time a node comes online.

Watcher implementations will now filter out Bookmark and Error action events
which don't represent Pod state changes. Future enhancement might be to use
bookmarks on reconnect to catchup on events.

Deprecated KubernetesLauncher `watcher` field which does not appear to be
used by the plugin anywhere and is a potential for threading issues when lots of
agents are launched at once.