Allow agent pods to be Pending while launching
This commit relaxes the requirement for Pods to be `Running` while the
agent launcher is waiting for the agent to connect. This should address
JENKINS-68649, which is caused by pods on virtual nodes still being in
the `Pending` for a short while once all of its containers are running.

Fixes: JENKINS-68649
Jira: JENKINS-68649