review: invert pod status check
Thic commit changes the pod status check that is performed while waiting
for a new agent to connect. Instead of checking that the pod status is
one of a set of 'valid' options, we ensure that the status is _not_ one
that implies that the pod has terminated.

The commit also changes how the wait is terminated when this condition
arises. Instead of breaking out of the wait loop and replying on the
final timeout check to notice that the slave has not connected, we
immediately throw an exception. This should prevent the `LAUNCH_FAILED`
metric from being double-counted, and also prevents the `TIMEOUT` metric
from being incremented, as this is _not_ a timeout.