[JENKINS-70217] Default ContainerTemplate command and arg to empty string when using casc
When defining a container template using Casc and unspecified command
and arg, it seems more natural that the default entrypoint and args are
executed. So setting these to empty string when constructing a new
instance seems more natural.

This doesn't change the behaviour of creating a new container template
using the UI, as the default value will come from jelly because the
ContainerTemplate instance will be null when populating the jelly
fragment.