Apply restricted PSS to all containers when enabled
When restrictedPssSecurityContext = true in a kubernetes cloud, all created containers will be augmented with a security context compatible with "restricted" Pod Security Admission.

If a security context has been manually set earlier, this won't be overridden. In such case, the pod will be rejected and the user should correct its pod definition.