Fetch Pod reference if unassigned
This change updates the KubernetesSlave to perform a lookup of the pod
when the pod reference is `null`. Usually the Pod reference is assigned at
launch. If Jenkins restarts and a Kubernetes agent attempts to reconnect this
reference is lost.

A new method `KubernetesCloud#getPodResource` was added to improve testability and encapsulate common lookup operations. Notably not all call
sites were checking for the `@Nullable` `namespace` field when calling
`.inNamespace()`, which does not accept null values.