Refactor ContainerTemplate command split function
This change delegates `PodTemplateBuilder` command line split functions to
a new `PodTemplateUtils#splitCommandLine` function. This new function
improves command line quote handling, including support for single quotes.

This new function was moved to `PodTemplateUtils` so it is in the same
class as `substituteEnv`, which it depends on, and so it can be used
extensions that might need to build a fabric8 model from a `ContainerTemplate`.