Test with Java 25 and Java 21
Java 25 released Sep 16, 2025.  The Jenkins project wants to support
Java 25 soon.  Assuring that plugins pass tests on Java 25 is part of
the effort to support Java 25.

Replaces Java 17 with Java 21 as the Java version for kind tests.
The parent pom configures the Java 21 compiler to generate Java 17
byte code.  We've not found any issues where byte code generated by
the Java 21 compiler resulted in bugs that were not also visible when
the code was compiled with Java 17.  Rather than spend the money to run
tests with Java 17, let's just run the tests with Java 21 and Java 25.

Replaces Java 21 with Java 25 for the tests that are run outside kind.

Testing done:

* Confirmed that tests run locally with Java 25
* Will confirm that the expected Java versions are used
* Will confirm that kind tests pass on ci.jenkins.io